feat: introduce MimeType enum for consistent media type handling

+++ lib/app/services/media_service/media_encryption_service.m.dart
@@ -10,12 +10,14 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/exceptions/exceptions.dart';
 import 'package:ion/app/features/core/model/media_type.dart';
+import 'package:ion/app/features/core/model/mime_type.dart';
 import 'package:ion/app/features/core/providers/ion_connect_media_url_fallback_provider.r.dart';
 import 'package:ion/app/features/ion_connect/model/media_attachment.dart';
 import 'package:ion/app/services/compressors/brotli_compressor.r.dart';
 import 'package:ion/app/services/file_cache/ion_file_cache_manager.r.dart';
 import 'package:ion/app/services/logger/logger.dart';
 import 'package:ion/app/services/media_service/media_service.m.dart';
+import 'package:mime/mime.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
@@ -75,12 +77,22 @@ class MediaEncryptionService {
         );
 
         final decryptedFileBytes = Uint8List.fromList(decryptedFileBytesList);
+        final decryptedFile = File.fromRawPath(decryptedFileBytes);
 
-        final fileExtension = attachment.mimeType.split('/').last;
+        final mimeType = lookupMimeType(decryptedFile.path, headerBytes: decryptedFileBytes);
+        if (mimeType == null) {
+          throw FailedToDecryptFileException();
+        }
+
+        final fileExtension = extensionFromMime(mimeType);
+        if (fileExtension == null) {
+          throw FailedToDecryptFileException();
+        }
 
         await fileCacheService.removeFile(url);
 
-        if (attachment.mediaType == MediaType.unknown) {
+        final mediaType = MediaType.fromMimeType(mimeType);
+        if (mediaType == MediaType.unknown) {
           final decompressedFile = await brotliCompressor.decompress(decryptedFileBytes);
 
           final decryptedFile = await fileCacheService.putFile(
@@ -156,7 +168,7 @@ class MediaEncryptionService {
           path: compressedEncryptedFile.path,
           width: mediaFile.width,
           height: mediaFile.height,
-          mimeType: mediaFile.mimeType,
+          mimeType: MimeType.generic.value,
           duration: mediaFile.duration,
         );
